Refactor to move most logic outside of the command file.
